requirements.txt
requests
plexapi
urllib3
python-dotenv